+++ TradeSkillMaster/ChangeLog.md
@@ -1,9 +1,21 @@
 [Welcome to TradeSkillMaster 4!](https://blog.tradeskillmaster.com/tradeskillmaster-4-released/)
 
 
-## v4.2.2 Changes
+## v4.3 Changes
 
-* Fixed sorting of the crafting queue and made it significantly less laggy to scroll through and load
-* Fixed bug with posting of caged pets
+* You can now shift-click on item links or items in your bags to easily search for them in the TSM profession UI
+* The TSM profession UI now supports searching by materials in addition to the crafted item name
+* A bunch of optimizations were made to sniper to make it much easier to buy items right as they appear as fast as possible
+* Updated milling and prospecting ratios
+* Greatly reduced FPS impact when TSM needs to rebuild its item database
+* Greatly reduced "Starting scan..." delay when first starting a scan after the item database was rebuilt
+* Changed the default Shopping operation to have a max price of "dbmarket" and to show auctions above max price (existing operations aren't affected)
+* Fixed bug with multiple Mailing operations not sending the correct quantities
+* Fixed issue with bidding through Shopping
+* Fixed a few enchants which TSM Crafting didn't properly support
+* Fixed bug with "Repost Higher Threshold" setting in Auctioning operations not saving
+* Fixed UI bug with "Set minimum profit?" toggle in Crafting operations
+* Fixed UI bug with "Restock selected groups" button in Crafting being disabled when groups are selected
+* Fixed error some users were getting when logging in
 
 [Known Issues](http://support.tradeskillmaster.com/display/KB/TSM4+Currently+Known+Issues)